实现问题回复-事务&实现问题回复-页面提交回复

+++ src/main/java/cn/edu/hncj/forum/dto/QuestionDTO.java
@@ -15,7 +15,7 @@ public class QuestionDTO {
     private Long gmtModified;
     private Integer viewCount;
     private Integer likeCount;
-    private Integer creator;
+    private Long creator;
     private Integer commentCount;
     private String tag;
     private User user;
